Moves JS source to correct directory
