+++ js/src/lib/cartodb_collection.js
