+++ js/src/lib/format.js
